Split processor tests into multiple files
